Pass the Google consent page.
PhantomJS can still render the consent page, so interacting with is not a problem; it's the Youtube page that comes after it.
Anyway, I needed PhantomJS to hit the Youtube page, if I want to prove it doesn't work there anymore; to that end I've left some comment-out lines that can take screenshots if needed.